test: move datastore to stack-viewer tests
